clean up homepage, fix some styling
